data inspection
I have inspected the data of station#95 that is the main geographical point where we are going to conduct the analysis. I have inspected the data relative to the quality of the air, wind and temp. I have subsequently created a new column that stores the mean value for pm25 of each day. (observations are hourly)